Fix: registrar cobro de saldo fallaba por ambigüedad de tipo en SQL
La suma $3 + $4 en marcarSaldoPagado no tenía tipo explícito, y Postgres
no podía resolver el operador (+) entre parámetros "unknown", rompiendo
todo intento de confirmar el cobro de un saldo pendiente. Se agregan
casts ::numeric explícitos.

De paso, se corrige el mensaje de error de login (la ñ sin codificar
rompía el encoding de la URL de redirect).

+++ src/lib/db.ts
@@ -204,8 +204,8 @@ export async function marcarSaldoPagado(
        saldo_socio = $4,
        cotizacion_saldo = $5,
        monto_luz = $6,
-       saldo_pendiente = $3 + $4,
-       monto = monto_yo + monto_socio + $3 + $4
+       saldo_pendiente = $3::numeric + $4::numeric,
+       monto = monto_yo + monto_socio + $3::numeric + $4::numeric
      WHERE id = $1`,
     [id, fecha, saldoYo, saldoSocio, cotizacionSaldo, montoLuz],
   );
